jerkins: init: Conditionally pull bazel-remote

+++ shell/init.sh
@@ -18,9 +18,10 @@ echo "----------------------------------------------------"
 echo "Pull Bazel-remote from github releases"
 echo "----------------------------------------------------"
 
-rm -rf bin/bazel-remote
-wget https://github.com/buchgr/bazel-remote/releases/download/v2.6.1/bazel-remote-2.6.1-linux-amd64 -O bin/bazel-remote
-chmod +x bin/bazel-remote
+if [ ! -f "bin/bazel-remote" ]; then
+    wget https://github.com/buchgr/bazel-remote/releases/download/v2.6.1/bazel-remote-2.6.1-linux-amd64 -O bin/bazel-remote
+    chmod +x bin/bazel-remote
+fi
 
 echo " "
 echo "----------------------------------------------------"
